+++ packages/gasket-nextjs-plugin/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+###
+
+- [#23] Dismantling core-plugin
+
 ### 1.1.1
 
 - Fix for missing dependency during create command
@@ -9,3 +13,5 @@
 - Renaming hooks and removing the `webpack` hook
 - Separate next plugins from core
 - Initial release.
+
+[#23]: https://github.com/godaddy/gasket/pull/23